requirements.txt
# FastMCP - Model Context Protocol framework
fastmcp>=2.9.0

# HTTP client for Bitbucket API
httpx>=0.27.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation and parsing
pydantic>=2.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
